[LEVWEB-71] Remove hover decoration on crown image
This removes the mock-hyperlink hover behaviour from the crown image's
parent div. This is desirable as we no longer have a link on this image.

"Design Review: Crown in Header is a link to nowhere

The Crown image in the header has a hover state so it looks like a link
but goes nowhere. Suggest removing this so it is just an image."