[LEVWEB-162] Log-out support
Adds a "Logout" button to all pages. The button links through to the
KeyCloak logout URL, which then redirects back to a new endpoint,
/logged-out.

Unfortunately this does not, on its own, revoke access to the website.
This is probably a bug in go-keycloak-proxy but we are not certain at
this time. In order to secure the logout the endpoint clears the
kc-access cookie. Once the bug is fixed in go-keycloak-proxy this extra
functionality should be removed.

Note: This is not fully tested, as I could not think of a way to do so
without first deploying.