Remove mock-kc-proxy
Removes mock-kc-proxy in favour of simply sending some extra headers on
requests to the API. This allows lev-web to be hooked up to the API
directly. In production these headers will be overwritten by the real
keycloak-proxy.

This is the same approach we use when acceptance testing the API with
dredd.