[LEVWEB-625] A less hideous refactor?
Hopefully a less hideous refactor of the tests for link checking.
Defines a function that takes a boolean for whether the pages you will
be checking are error pages or not and returns a faction that will check
a page of that type. (That you can simply map over an array of pages.)